% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_biorad2.R
\name{load_biorad2}
\alias{load_biorad2}
\title{Load a qPCR file from a BioRad2 machine}
\usage{
load_biorad2(datafile)
}
\arguments{
\item{datafile}{A comma delimited (.csv) file containing data produced by the Bio-Rad Opticon Monitor 3 software.}
}
\value{
Invisibly, creates a dataframe for subsequest processing of qPCR data.
}
\description{
Imports data from Bio-Rad Opticon qPCR machines that utilize the Bio-Rad Opticon Monitor 3 software. This function only accepts raw data from the program, without any calculations on user-input well sets.
}
\examples{
load_biorad2("my_data.csv")
load_biorad2("./data/my_data.csv")
}
