% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_bioradcfx.R
\name{load_bioradcfx}
\alias{load_bioradcfx}
\title{Load a qPCR file from a BioRad CFX 96 well machine}
\usage{
load_bioradcfx(file)
}
\arguments{
\item{the}{file Path in quotes}
}
\value{
A tibble table of the input file. This inlcudes columns with the sample designation, starting quantity (copy number), mean, and standard deviation.
}
\description{
This function loads a .csv file as a tibble table. You must first convert your .xlsx file into a .csx file before using. It assumes that the first row of each column
contains the column names and that each row corresponds to a well run. It also assumes that the format of the .csv file has the following column names:
Well, Fluor, Target, Content, Sample, Biological Set Name,	Cq,	Cq Mean,	Cq Std. Dev,	Starting Quantity (SQ),
Log Starting Quantity	SQ, Mean	SQ, Std. Dev,	Set Point,	Well Note. If these are not the header names, the file may not
download properly.
}
