% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_merge.R
\name{meta_merge}
\alias{meta_merge}
\title{Combine qPCR and metadata}
\usage{
meta_merge(file, df)
}
\arguments{
\item{file}{a specified comma delimited file containing metadata associated with qPCR Sample data}

\item{df}{a data frame including qPCR data}
}
\value{
A merged data frame including combined qPCR and metadata
}
\description{
This function loads comma delimited file containing metadata associated with qPCR data
and combines it with qPCR data by the column Sample. The Sample names and column header
"Sample" must be the same in both combined data sets.
}
