% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_biom.R
\name{norm_biom}
\alias{norm_biom}
\title{Normalize by extraction elution volume and biomass input}
\usage{
norm_biom(df, x = 1, y = 1, z)
}
\arguments{
\item{df}{a data frame resulting from either norm_pd() or norm_sd()}

\item{x}{a numeric value, equal to elution volume}

\item{y}{a numeric value, equal to biomass input}

\item{z}{a new column name; include in quotes}
}
\value{
A data frame including Sample, Mean.Copies, and SD after correcting for total elution volume and biomass input.
}
\description{
This function takes a data frame and normalizes Mean.Copies.ul by the quantity of biomass in
the original extract. The function multiplies mean copies by the volume of the final elutant
following DNA extraction. For example, if an extraction results in a final elution volume of
100 microliters, x would equal 100 and the resulting normalization would represent the mean
copies of DNA in the elution volume. y can then be used to normalize copy number for a particular
unit of mass or volume of the input material for the DNA extraction. For example, if DNA was
extracted from 0.25g of soil, y would equal 0.25, and the resulting normalization would be
mean copies per gram of soil; Z is used to define these final units and name the resulting column.
This function should only be used after norm_pd or norm_sd.
}
\examples{
norm_biom(quant2, 100, 0.25, "Mean.Copies/g")
}
