% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_biom2.R
\name{norm_biom2}
\alias{norm_biom2}
\title{Normalize by extraction elution volume and biomass input if different masses per sample}
\usage{
norm_biom2(df, x, y, z)
}
\arguments{
\item{df}{a data frame resulting from either norm_pd(), norm_sd(), or meta_merge()}

\item{x}{a numeric value, equal to elution volume}

\item{y}{a new column name; include in quotes}

\item{z}{a column including mass information}
}
\value{
A data frame including Sample, Mean.Copies per biomass input, and SD after correcting for total elution volume and biomass input.
}
\description{
This function takes a data frame and normalizes Mean.Copies.ul by the quantity of biomass in
the original extract. The function multiplies mean copies by the volume of the final elutant
following DNA extraction (x). For example, if an extraction results in a final elution volume of
100 microliters, x would equal 100 and the resulting normalization would represent the mean
copies of DNA in the elution volume. df[[z]] can then be used to normalize copy number for a particular
unit of mass or volume of the input material for the DNA extraction. This is to be used if
each sample has a different mass of initial input. Z is used to define these final units and name the resulting column.
This function should only be used after norm_pd or norm_sd and meta_merge().
}
\examples{
norm_biom(quant4, 100, "Mean.Copies.gdw", "Soil.dw.")
}
