% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_gcopy.R
\name{norm_gcopy}
\alias{norm_gcopy}
\title{Normalize 16SrRNA qPCR data by gene copy number to obtain cells/mL}
\usage{
norm_gcopy(df, x = 1, y = 1, z = 1, n = 1)
}
\arguments{
\item{df}{The input dataframe}

\item{x}{Elution volume of the extracted DNA in microliters}

\item{y}{Template volume of DNA in the PCR reaction in microliters}

\item{z}{Sample volume subjected to DNA extraction in milliliters}

\item{n}{Number of 16SrRNA gene copies in the genome}
}
\value{
A dataframe including Sample, Mean.Conc, SD.Conc
}
\description{
This function uses a four-column dataframe containing "Samples", "Copies",
"Mean", and "SD" that has already been processed to remove outliers.
It normalizes "Mean.Copies" by the organism's gene copy number to provide a 
cell concentration (cells/ml). This fuction is applicable for axenic cultures.
}
\details{
This method derives from Ritalahti et al. 2006 in AEM, using the equation: 
i) Gene Copies per mL = ((gene copies per reaction mix) x (volume of DNA[ul]) / (3 ul DNA per reaction mix) x (ml sample used))
ii) Cells per mL = Gene Copies per mL / Gene Copies per Genome
}
\examples{
norm_gcopy(yourfile, 50, 2, 1.8, 2)

}
