% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_pd.R
\name{norm_pd}
\alias{norm_pd}
\title{Normalize by Pipette volume and dilution factor}
\usage{
norm_pd(df, x = 1, y = 1)
}
\arguments{
\item{df}{a data frame}

\item{x}{a numeric value}

\item{y}{a numeric value}
}
\value{
A data frame including Sample, Mean.Copies.ul, and SD after correcting for pippet volume and dilutions.
}
\description{
This function takes a data frame and normalizes Mean.Copies by the amount of extract
used in each pcr reaction and the dilution factor. This function can only be used following
repsout(). Both the pipetted quantity (x) and the dilution factor (y) are user defined and
the default is set to 1 (indicating no dilution). The dilution factor argument should not be
modified if only a subset of samples were diluted; instead, use the function norm_sd() prior
to this function.
}
\examples{
norm_pd(quant2, 2, 10)
}
