% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cellconc.R
\name{plot_cellconc}
\alias{plot_cellconc}
\title{Plots cell concentration from qPCR data.  This should only be used after meta_merg() that has
added a column for Organism and Time, resulting in five total columns.}
\usage{
plot_cellconc(df, a, b)
}
\arguments{
\item{df}{a dataframe with five columns}

\item{a}{title for the graph in quotes}

\item{b}{title for x axis in quotes, default is Time(hrs) if not defined}
}
\value{
A plot of cell concentration versus time, faceted by organism
}
\description{
This function plots a numeric value (cells/mL) by sample from a specified data frame.
}
