% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pcr.R
\name{plot_pcr}
\alias{plot_pcr}
\title{Plots qPCR data}
\usage{
plot_pcr(df, i, j, k)
}
\arguments{
\item{df}{a data frame}

\item{i}{a vector (column) of type character or factor belonging to df}

\item{j}{a vector (column) of type numeric belonging to df}

\item{k}{a vector (column) of type character or factor belonging to df}
}
\value{
A plot of copies by sample; includes error bars with the standard deviation
}
\description{
This function plots a numeric value (e.g. Mean.copies) by sample from a specified data frame.
If used after meta_merge(), it also allows you to specify a grouping variable (k) to compare
types of samples by color.
}
