% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subdil.R
\name{subdil}
\alias{subdil}
\title{Normalize a subset of samples by the dilution factor}
\usage{
subdil(df, row.names = x, y = 1)
}
\arguments{
\item{df}{a data frame}

\item{y}{a numeric value}

\item{x}{a numeric value}
}
\value{
A data frame including Sample, Mean.Copies, and SD after correcting for dilutions.
}
\description{
This function takes a data frame and normalizes a subset of Mean.Copies by the dilution factor.
The dilution factor (y) is user defined and the default is set to 1 (indicating no dilution).
A vector including Samples that were diluted needs to be created prior to using this function.
}
\examples{
x<-c('1101-0001|1101-0002|1101-0003')
subdil(quant1, x, 10)
}
