% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_quantmasterviaa7.R
\name{load_quantmasterviia7}
\alias{load_quantmasterviia7}
\title{load_quantmasterviia7}
\usage{
load_quantmasterviia7(file)
}
\arguments{
\item{file}{The qPCR output (.csv) from a QuantMasterViiA7 Instrument}
}
\value{
Returns a tibble for next step, normalization
}
\description{
Imports raw qPCR data from a QuantMaster Viaa7 Instrument, converts it to a smaller tibble containing Sample Name and Quantity
}
\examples{
load_quantmasterviia7("jenny2_qPCR.csv")

}
